Move workspace tools into tools/workspace
