+++ README.md
@@ -119,7 +119,7 @@ libraries of type `assets` and `styles`**._
 To generate a new library, run the below command\* and follow the instructions:
 
 ```sh
-bun nx g workspace-tools:library-generator
+bun nx g @energinet-datahub/tools/workspace:library
 ```
 
 While rarely needed, it is also possible to generate an entirely new domain.
@@ -127,7 +127,7 @@ Running the following command\* will create a new domain with
 `data-access-api`, `feature`, and `shell` libraries included:
 
 ```sh
-bun nx g workspace-tools:domain-generator
+bun nx g @energinet-datahub/tools/workspace:domain
 ```
 
 > Also available in [Nx Console](https://nx.dev/core-features/integrate-with-editors).
