+++ tools/workspace/README.md
@@ -1,0 +1,25 @@
+# Tools for generating code
+
+To run the translation keys executor:
+
+```sh
+bun nx run @energinet-datahub/tools/workspace:translation-keys
+```
+
+To run the domain generator:
+
+```sh
+bun nx run @energinet-datahub/tools/workspace:domain
+```
+
+To run the library generator:
+
+```sh
+bun nx run @energinet-datahub/tools/workspace:library
+```
+
+To run the watt component generator:
+
+```sh
+bun nx run @energinet-datahub/tools/workspace:watt-component
+```
