+++ tools/workspace/executors.json
@@ -3,7 +3,7 @@
     "generate-translation-keys": {
       "implementation": "./src/executors/generate-translation-keys/executor",
       "schema": "./src/executors/generate-translation-keys/schema.json",
-      "description": "generate-translation-keys executor"
+      "description": "Generate translation keys executor"
     }
   }
 }
