+++ tools/workspace/src/generators/library/files/configuration/test-setup/test-setup.ts__tmpl__
