+++ tools/workspace/src/generators/library/files/data-access/store/__product__-__domain__-__type__-__name__.store.ts__tmpl__
