+++ tools/workspace/src/generators/library/files/environments/test-setup/test-setup.ts__tmpl__
