+++ tools/workspace/src/generators/library/files/shell/module/__product__-__domain__-shell.routes.ts__tmpl__
