+++ tools/workspace/src/generators/library/files/shell/test-setup/test-setup.ts__tmpl__
