+++ tools/workspace/src/generators/library/schema.json
@@ -1,7 +1,7 @@
 {
   "$schema": "http://json-schema.org/schema",
   "cli": "nx",
-  "$id": "library-generator",
+  "$id": "library",
   "description": "Generate a library that is going to be used in DataHub.",
   "type": "object",
   "properties": {
