+++ tools/workspace/src/generators/watt-component/files/src/index.ts.template
