+++ tools/workspace/src/generators/watt-component/files/src/watt-__fileName__.component.spec.ts.template
