+++ tools/workspace/src/generators/watt-component/files/src/watt-__fileName__.component.ts.template
