+++ tools/workspace/src/generators/watt-component/files/src/watt-__fileName__.stories.ts.template
