+++ tools/workspace/src/generators/watt-component/generator.ts
