+++ tools/workspace/src/generators/watt-component/schema.d.ts
