--- workspace-tools/README.md
@@ -1,11 +1,0 @@
-# workspace-tools
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build workspace-tools` to build the library.
-
-## Running unit tests
-
-Run `nx test workspace-tools` to execute the unit tests via [Jest](https://jestjs.io).
