--- workspace-tools/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "workspace-tools",
-  "version": "0.0.1",
-  "dependencies": {
-    "ts-morph": "21.0.1",
-    "prettier": "3.3.3",
-    "@nx/devkit": "20.4.5",
-    "@nx/angular": "20.4.5",
-    "tslib": "2.4.0"
-  },
-  "type": "commonjs",
-  "main": "./src/index.js",
-  "typings": "./src/index.d.ts",
-  "generators": "./generators.json",
-  "executors": "./executors.json"
-}
