--- workspace-tools/project.json
@@ -1,52 +1,0 @@
-{
-  "name": "workspace-tools",
-  "$schema": "../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "workspace-tools/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/workspace-tools",
-        "main": "workspace-tools/src/index.ts",
-        "tsConfig": "workspace-tools/tsconfig.lib.json",
-        "assets": [
-          "workspace-tools/*.md",
-          {
-            "input": "./workspace-tools/src",
-            "glob": "**/!(*.ts)",
-            "output": "./src"
-          },
-          {
-            "input": "./workspace-tools/src",
-            "glob": "**/*.d.ts",
-            "output": "./src"
-          },
-          {
-            "input": "./workspace-tools",
-            "glob": "generators.json",
-            "output": "."
-          },
-          {
-            "input": "./workspace-tools",
-            "glob": "executors.json",
-            "output": "."
-          }
-        ]
-      }
-    },
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "outputs": ["{options.outputFile}"]
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "workspace-tools/jest.config.ts"
-      }
-    }
-  },
-  "tags": []
-}
